Revert r26584: "Upgrade to Perst 4.05"
This is causing more problem then it supposed to fix.

git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/XMLSpider@26646 67a373e5-eb02-0410-a15c-ee090a768436

+++ org/garret/perst/impl/ReplicationDynamicSlaveStorageImpl.java
@@ -1,10 +1,9 @@
 package plugins.XMLSpider.org.garret.perst.impl;
 
-import java.io.IOException;
-import java.net.Socket;
+import java.io.*;
+import java.net.*;
 
-import plugins.XMLSpider.org.garret.perst.IFile;
-import plugins.XMLSpider.org.garret.perst.StorageError;
+import plugins.XMLSpider.org.garret.perst.*;
 
 
 public class ReplicationDynamicSlaveStorageImpl extends ReplicationSlaveStorageImpl
@@ -14,7 +13,7 @@ public class ReplicationDynamicSlaveStorageImpl extends ReplicationSlaveStorageI
         this.port = port;
     }
 
-    public void open(IFile file, long pagePoolSize) {
+    public void open(IFile file, int pagePoolSize) {
         initialized = false;
         prevIndex = -1;
         outOfSync = true;
